+++ changelog.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## V 0.1.2
+1. Revise dependencies
+2. Add currency details look up from currency code
+3. Add intitial testing 
+4. Patch vunerabilities
+5. Add currencyCode ruralCurrency function which taks a code such as GBP and returns an object of currency details. 
+
 ## V 0.1.11
 1. Removed lodash as a hard dependency and utilize Object.entries and a for loop instead of the map requirement.
 2. Start revising for the V 0.2 data process.